feat: Improve app stability and user experience
This commit introduces several enhancements:

- **Web App Stability:** Implemented a referrer policy in `index.html` to prevent referrer leakage, enhancing privacy and security.
- **Notification Service:** Refined the logic for subscribing and unsubscribing users from push notifications in `Layout.tsx`, ensuring a more robust user experience.
- **UI/UX Improvements:**
    - Clarified button text from "ANALISA INFORMASI OBAT" to "ANALISA OBAT" in `MedicationView.tsx`.
    - Shortened the month display in `MembersList.tsx` for better date readability.
    - Enhanced the WA link generation in `ContactsView.tsx` to handle various phone number formats more gracefully.
- **Backend Service (`spreadsheetService.ts`):**
    - Updated the version of the spreadsheet service to v8.8.
    - Introduced a new constant `SOURCE_SHEET_ID` for better data source management.
    - Implemented a more robust `transformDriveUrl` function to handle different Google Drive URL formats, including ensuring it correctly transforms URLs for image display and prevents re-processing already formatted URLs.
- **Geolocation:**
    - Improved the `ContactsView.tsx` component's geolocation handling:
        - Optimized the `startLocationWatch` function by cleaning up existing watches before starting a new one.
        - Refined the geolocation options to use a timeout strategy that prioritizes high accuracy when needed and falls back to lower accuracy more quickly when appropriate.
        - Removed outdated comments about geolocation strategy.

These changes collectively contribute to a more stable, secure, and user-friendly application.

+++ components/MembersList.tsx
@@ -30,7 +30,7 @@ const formatDateUpper = (dateString: string) => {
   if (!dateString) return '-';
   const d = new Date(dateString);
   if (isNaN(d.getTime())) return dateString;
-  return d.toLocaleDateString('id-ID', { day: '2-digit', month: 'long', year: 'numeric' }).toUpperCase();
+  return d.toLocaleDateString('id-ID', { day: '2-digit', month: 'short', year: 'numeric' }).toUpperCase();
 };
 
 const MembersList: React.FC<MembersListProps> = ({ members, language, onAddMember, onUpdateMember, onDeleteMember, onSelectMember, selectedId }) => {
